1. 改进download-release.sh: 避免重复拷贝压缩文件 2. 修改ubuntu/util.sh: 修改下载easy-rsa.tar.gz的行为 -> 从下载文件夹拷贝 3. 增加get-kubernetes-deps.sh: 获取安装kubernetes所需的easy-rsa.tar.gz

+++ cluster/ubuntu/util.sh
@@ -365,8 +365,14 @@ function kube-up() {
   # downloading tarball release
   "${KUBE_ROOT}/cluster/ubuntu/download-release.sh"
 
-  # Fetch the hacked easyrsa that make-ca-cert.sh will use
-  curl -L -O https://storage.googleapis.com/kubernetes-release/easy-rsa/easy-rsa.tar.gz > /dev/null 2>&1
+  if [ ! -f easy-rsa.tar.gz ]; then
+    if [ ! -f ${DOWNLOAD_PATH}/easy-rsa.tar.gz ]; then
+      # Fetch the hacked easyrsa that make-ca-cert.sh will use
+      curl -L -O https://storage.googleapis.com/kubernetes-release/easy-rsa/easy-rsa.tar.gz > /dev/null 2>&1
+    else
+      cp ${DOWNLOAD_PATH}/kubernetes/easy-rsa.tar.gz .
+    fi
+  fi
 
   if ! check-CNI-config; then
     return
